@@ -2,50 +2,35 @@
   "name": "realtime",
   "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@reduxjs/toolkit": "^1.8.6",
-    "@tailwindcss/forms": "^0.5.3",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^13.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "moment": "^2.29.4",
-    "pocketbase": "^0.7.1",
-    "react": "^18.2.0",
-    "react-debounce-input": "^3.3.0",
-    "react-dom": "^18.2.0",
-    "react-redux": "^8.0.4",
-    "react-scripts": "5.0.1",
-    "react-toastify": "^9.0.8",
-    "redux-persist": "^6.0.0",
-    "web-vitals": "^2.1.0"
-  },
+  "type": "module",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "@reduxjs/toolkit": "^2.3.0",
+    "@tailwindcss/forms": "^0.5.9",
+    "@vitejs/plugin-react-swc": "^3.7.1",
+    "moment": "^2.30.1",
+    "pocketbase": "^0.21.5",
+    "react": "^18.3.1",
+    "react-debounce-input": "^3.3.0",
+    "react-dom": "^18.3.1",
+    "react-redux": "^9.1.2",
+    "react-toastify": "^10.0.6",
+    "redux-persist": "^6.0.0"
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.12",
-    "postcss": "^8.4.17",
-    "tailwindcss": "^3.1.8"
+    "@types/node": "^22.8.6",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@typescript-eslint/eslint-plugin": "^8.12.2",
+    "@typescript-eslint/parser": "^8.12.2",
+    "autoprefixer": "^10.4.20",
+    "postcss": "^8.4.47",
+    "tailwindcss": "^3.4.14",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.10"
   }
 }
